Implemented transfer validation chain of responsability

+++ src/main/java/com/paulo/minipicpay/infrastructure/validate/BaseValidationHandler.java
@@ -3,6 +3,8 @@ package com.paulo.minipicpay.infrastructure.validate;
 import com.paulo.minipicpay.model.User;
 import com.paulo.minipicpay.model.ValidateTransferHandler;
 
+import java.math.BigDecimal;
+
 public abstract class BaseValidationHandler implements ValidateTransferHandler {
 
     private final ValidateTransferHandler next;
@@ -11,7 +13,7 @@ public abstract class BaseValidationHandler implements ValidateTransferHandler {
         this.next = next;
     }
 
-    protected void next(User user, Double amount) {
+    protected void next(User user, BigDecimal amount) {
         next.validate(user, amount);
     }
 }
